JAMES-2645 Refactor cassandra RRTs by adding a proper DAO class with basic tests
